Crate pyo3cls

This crate declares only the proc macro attributes, as a crate defining proc macro attributes must not contain any other public items.

Attribute Macros

pyclass

pyfunction

pymethods

pymodule
Internally, this proc macro create a new c function called PyInit_{my_module} that then calls the init function you provided

pyproto