Crate pyo3cls

This crate declares only the proc macro attributes, as a crate defining proc macro attributes must not contain any other public items

Functions

class

function

methods

mod2init

mod3init

proto